chore(builder): removed header and footer components from builder and made them in Playground
