--- packages/builder/base/src/header/index.ts
@@ -1,3 +1,0 @@
-export { PreviewButton } from "./PreviewButton";
-export { Header } from "./Header";
-export { Logo } from "./Logo";
