+++ packages/builder/base/src/index.ts
@@ -1,7 +1,5 @@
 export { Canvas } from "./Canvas";
 export { Container } from "./Container";
-export { Footer } from "./Footer";
-export { Header, PreviewButton, Logo } from "./header";
 export { Settings } from "./Settings";
 export { Workspace } from "./Workspace";
 export { Sidebar, FloatingSidebar, SidebarItem } from "./sidebar";
